+++ src/connection_factory.rs
@@ -1,5 +1,7 @@
+use std::borrow::Cow;
 use std::error::Error as StdError;
 use std::fmt;
+use std::sync::Arc;
 
 use futures::{future::Future, Poll};
 
@@ -8,9 +10,10 @@ use crate::Poolable;
 pub trait ConnectionFactory {
     type Connection: Poolable;
     fn create_connection(&self) -> NewConnection<Self::Connection>;
-    fn connecting_to(&self) -> &str;
+    fn connecting_to(&self) -> Cow<[Arc<String>]>;
 }
 
+#[derive(Debug)]
 pub struct NewConnectionError {
     cause: Box<dyn StdError + Send + 'static>,
 }
@@ -29,6 +32,16 @@ impl NewConnectionError {
 impl fmt::Display for NewConnectionError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "could not create a new connection: {}", self.cause)
+    }
+}
+
+impl StdError for NewConnectionError {
+    fn description(&self) -> &str {
+        "could not create a new connection"
+    }
+
+    fn cause(&self) -> Option<&dyn StdError> {
+        Some(&*self.cause)
     }
 }